tests/wpt/tests/css/css-align/parsing/grid-column-gap-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: parsing grid-column-gap with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-align/#grid-gap-legacy">
<link rel="author" title="Takuya Kurimoto" href="mailto:takuya004869@gmail.com">
<meta name="assert" content="grid-column-gap supports only the grammar '<length-percentage> | normal'.">
<meta name="assert" content="grid-column-gap rejects negative <length-percentage>.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("grid-column-gap", "auto");

test_invalid_value("grid-column-gap", "10");
test_invalid_value("grid-column-gap", "10px 20px");
test_invalid_value("grid-column-gap", "-1px");
test_invalid_value("grid-column-gap", "-10%");
</script>
</body>
</html>
